+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next version
 + Add Refresh Token Entity + Refresh Token Repository
++ Add UUID for token
 
 ## 0.11.0
 + Remove unused imports