+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Add Refresh Token Entity + Refresh Token Repository
 + Add UUID for token
++ Add Instant to TimeSource
 
 ## 0.11.0
 + Remove unused imports