+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Add Refresh Token Entity + Refresh Token Repository
 + Add UUID for token
 + Add Instant to TimeSource
++ Add Refresh Token Service
 
 ## 0.11.0
 + Remove unused imports