+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# 0.11.0
+# Next version
++ Generate token from token
+
+## 0.11.0
 + Remove unused imports
 + Fix "Type safety: The expression of type Page needs unchecked conversion to conform to Page<BankAccountTransactionsEntity>"
 + Return JSON if not authorized